diff_clean_up

+++ resip/stack/ConnectionBase.cxx
@@ -148,7 +148,7 @@ ConnectionBase::preparseNewBytes(int bytesRead, Fifo<TransactionMessage>& fifo)
       }
       case ReadingHeaders:
       {
-         unsigned int chunkLength = mBufferPos + bytesRead;
+         unsigned int chunkLength = static_cast<unsigned int>(mBufferPos + bytesRead);
          char *unprocessedCharPtr;
          MsgHeaderScanner::ScanChunkResult scanChunkResult =
             mMsgHeaderScanner.scanChunk(mBuffer,
@@ -168,13 +168,13 @@ ConnectionBase::preparseNewBytes(int bytesRead, Fifo<TransactionMessage>& fifo)
          }
 
          unsigned int numUnprocessedChars =
-            (mBuffer + chunkLength) - unprocessedCharPtr;
+             static_cast<unsigned int>(mBuffer + chunkLength - unprocessedCharPtr);
 
          if(numUnprocessedChars==chunkLength)
          {
             // .bwc. MsgHeaderScanner wasn't able to parse anything useful;
             // don't bother mMessage yet, but make more room in mBuffer.
-            size_t size = numUnprocessedChars*3/2;
+            unsigned int size = numUnprocessedChars*3/2;
             if (size < ConnectionBase::ChunkSize)
             {
                size = ConnectionBase::ChunkSize;
@@ -226,7 +226,7 @@ ConnectionBase::preparseNewBytes(int bytesRead, Fifo<TransactionMessage>& fifo)
             else
             {
                // ...but some of the chunk must be shifted into the next one.
-               size_t size = numUnprocessedChars*3/2;
+               unsigned int size = numUnprocessedChars*3/2;
                if (size < ConnectionBase::ChunkSize)
                {
                   size = ConnectionBase::ChunkSize;
@@ -251,7 +251,7 @@ ConnectionBase::preparseNewBytes(int bytesRead, Fifo<TransactionMessage>& fifo)
          }
          else
          {         
-            size_t contentLength = 0;
+            unsigned int contentLength = 0;
             
             try
             {
@@ -293,8 +293,8 @@ ConnectionBase::preparseNewBytes(int bytesRead, Fifo<TransactionMessage>& fifo)
             {
                // The message body is incomplete.
                DebugLog(<< "partial body received");
-               int newSize=resipMin(resipMax((size_t)numUnprocessedChars*3/2,
-                                             (size_t)ConnectionBase::ChunkSize),
+               int newSize=resipMin(resipMax(numUnprocessedChars*3/2,
+                                             (unsigned int)ConnectionBase::ChunkSize),
                                     contentLength);
                char* newBuffer = MsgHeaderScanner::allocateBuffer(newSize);
                memcpy(newBuffer, unprocessedCharPtr, numUnprocessedChars);
@@ -316,7 +316,7 @@ ConnectionBase::preparseNewBytes(int bytesRead, Fifo<TransactionMessage>& fifo)
                if (overHang > 0) 
                {
                   // The next message has been partially read.
-                  size_t size = overHang*3/2;
+                  unsigned int size = overHang*3/2;
                   if (size < ConnectionBase::ChunkSize)
                   {
                      size = ConnectionBase::ChunkSize;
@@ -360,7 +360,7 @@ ConnectionBase::preparseNewBytes(int bytesRead, Fifo<TransactionMessage>& fifo)
       }
       case PartialBody:
       {
-         size_t contentLength = 0;
+         UInt32 contentLength = 0;
 
          try
          {
@@ -397,6 +397,7 @@ ConnectionBase::preparseNewBytes(int bytesRead, Fifo<TransactionMessage>& fifo)
 
                Transport::stampReceived(mMessage);
                fifo.add(mMessage);
+							
                mMessage = 0;
             }
             mConnState = NewMessage;
