+++ rutil/WinCompat.cxx
@@ -16,10 +16,10 @@ static char* ConvertLPWSTRToLPSTR(LPWSTR lpwszStrIn)
    LPSTR pszOut = NULL;
    if (lpwszStrIn != NULL)
    {
-      int nInputStrLen = wcslen (lpwszStrIn);
+      size_t nInputStrLen = wcslen (lpwszStrIn);
 
       // Double NULL Termination
-      int nOutputStrLen = WideCharToMultiByte (CP_ACP, 0, lpwszStrIn, nInputStrLen, NULL, 0, 0, 0) + 2;
+      size_t nOutputStrLen = WideCharToMultiByte (CP_ACP, 0, lpwszStrIn, nInputStrLen, NULL, 0, 0, 0) + 2;
       pszOut = new char [nOutputStrLen];
 
       if (pszOut)
